+++ src/snowflake/cli/api/project/schemas/updatable_model.py
@@ -16,8 +16,7 @@ from __future__ import annotations
 
 from typing import Any, Dict
 
-from pydantic import BaseModel, ConfigDict, Field, ValidationError
-from snowflake.cli.api.project.errors import SchemaValidationError
+from pydantic import BaseModel, ConfigDict, Field
 from snowflake.cli.api.project.util import IDENTIFIER_NO_LENGTH
 
 
@@ -25,10 +24,7 @@ class UpdatableModel(BaseModel):
     model_config = ConfigDict(validate_assignment=True, extra="forbid")
 
     def __init__(self, *args, **kwargs):
-        try:
-            super().__init__(**kwargs)
-        except ValidationError as e:
-            raise SchemaValidationError(e)
+        super().__init__(**kwargs)
 
     def update_from_dict(self, update_values: Dict[str, Any]):
         """